рефакторинг Modal
