+++ src/shared/hocs/index.ts
@@ -1,6 +1,5 @@
 export { default as Droparea } from './Droparea.vue'
-export { default as ModalLayout } from './ModalLayout.vue'
 export { default as PopupTransition } from './PopupTransition.vue'
 export { default as Popover } from './Popover.vue'
-export { default as Modal } from './Modal.vue'
+export { default as Modal } from './Modal/Modal.vue'
 export { default as BackdropTransition } from './BackdropTransition.vue'
